requirements.txt
streamlit
chatterbox-tts
torch
torchaudio
pydub
librosa
numpy
